+++ registration.php
@@ -17,6 +17,6 @@
 
 \Magento\Framework\Component\ComponentRegistrar::register(
     \Magento\Framework\Component\ComponentRegistrar::MODULE,
-    'Blackbird_HyvaEstimateTimeShipping',
+    'Blackbird_EstimateTimeShippingHyvaCompatibility',
     __DIR__
 );