+++ changelog.md
@@ -1,4 +1,10 @@
 ## Version 1.0.0
+
+*Release the 2025-09-10*
+
+* **Fix** Correctly rename the namespace throughout the codebase to 'EstimateTimeShippingHyvaCompatibility'. This fix the problem with the not found plugin class.
+
+
 *Release the 2025-08-08*
 
 * **Compatibility** add Estimate Time Shipping module compatibility with Hyvä Themes